chore(vats): use new swingset type
Swingset's "VatAdminDevice" is now a "raw device", which parses the
`dispatch.deliver()` arguments to figure out what method is being invoked, so
it does not have a standard `buildRootDeviceNode` to glean type information
from. To compensate, I made it define a `VatAdminRootDeviceNode` type that
describes the effective API.

This changes the deduced types in packages/vats/ to use that type instead of
trying for the non-existent old one.