fix(swingset): add stub vat-upgrade calls (always fails)
This implements E(admin).upgrade(bundlecap, vatParameters), which queues the
request to the kernel, and awaits the response. The kernel always rejects the
request. The added unit test asserts that it gets an error of "unimplemented"
rather than some other error.

The next step is to enhance vat-warehouse to stop any current worker, replace
the source bundle, start a new worker, perform the upgrade delivery, make
sure it succeeds, and commit the results (or abort the changes).

refs #1848