fix(swingset): perform comms init during startVat, reduce DB reads
The comms vat uses `setup()` and therefore does not use a transcript. It does
not keep any state in RAM, which would cause non-determinism across process
restarts. It needs to initialize its DB state once, the first time the kernel
starts.

Before `disptach.startVat()` was called in `setup()`-based vats, it had no
clear signal for when this happened, so instead it had to perform a DB read
on every delivery, forever, just in case this was the very first one, and
initialization needed to happen. (In practice it only did the check durin
`dispatch.deliver()`, because we knew that would happen before any others).

Now that `startVat` is sent to `setup()`-based vats too, we can perform this
initialization at the beginning, and stop checking the DB on every subsequent
delivery.

This also removes an unnecessary assertion in comms `makeRemote()`, which
cost an additional DB read on every delivery that touched a remote
system (which was most of them).

An overly-sensitive virtual-objects unit test was fixed.

closes #4637