fix(swingset):initializeSwingset option to disable standard vats
With this change,

```
const initializationOptions = { addComms: false };
initializeSwingset(config, argv, hostStorage, initializationOptions);
```

will disable the creation of comms. `addVatAdmin`, `addVattp`, and `addTimer`
are also recognized. This is mostly useful for unit tests that want to ignore
the startVat events or VatID allocations made for these vats (or just want to
run marginally faster).