chore: weaken swingset dependency on vat-data to dev-dependency
Swingset *tests* use `@agoric/vat-data`, but not `src/`, so this
changes swingset/package.json to list it in `devDependencies`, not
`dependencies`.

Meanwhile, `@agoric/vat-data` lists swingset in its `devDependencies`
for its own tests. This gives us a cycle of `devDependencies` but not
normal dependencies.

Code which imports vat-data must (generally) be running inside a vat,
which means the application that decided to put that code into a vat
probably depends upon swingset: it had to import
`makeSwingsetController` to create the housing. But the vat-data
-importing code doesn't know that, and doesn't need a direct
dependency on swingset, because it never imports anything from
swingset itself.