refactor!: atomically update multiple parameters via governance (#5063)
* refactor!: atomically update multiple parameters via governance

The description of the change in the issue becomes  a set of
parameter changes so that if the vote passes, a single call from
governParam to the paramManager will make all the changes
atomically. If the update fails, all the changes will be unwound.

type declarations were updated.

Invitations require a two-phase update in paramManager, so all setters
had to be changed similarly for compatibility.

`setters` are no longer shared in the paramManager, though
`updateParams` still is, so the powerful facet of paramManager
continues to need to be closely held.

Tests for not updating partially are included

* refactor: simplify parameter description to a record from a list

* refactor: simplifiy updateParams with .map() and .fromEntries()

* refactor: make update methods more parallel

* chore: type repairs

* more type fixes

* clean up two-phase setInvitation

* use optional chaining

* fix: more optional chaining

* chore: correct param key for vaults

Co-authored-by: Turadg Aleahmad <turadg@agoric.com>