test(swingset): improve tests on result promises that appear in args
This doesn't change any behavior, but asserts that we're more tolerant
of messages whose arguments reference their own result
promise. Liveslots cannot generate these, and the kernel refuses to
accept them from non-pipelining vats, but remote systems might
construct such messages. So comms accept them, the kernel must not
kill comms when they appear, and they must be translatable and
deliverable to other vats. And liveslots must not choke when they
arrive.

The original concern of #5189 is a non-issue for liveslots-based vats,
because HandledPromise and E automatically harden the arguments before
the caller can mutate them. But this provides test coverage beyond
that point.

closes #5189