build(run-protocol): deploy-contracts script
Maintain E(home.scratch).get('installCache')
as a CopyMap<Hash,
            {
              sha512: string,
              installation: Installation,
              path?: string
            }>

Also, save installCache in the board.

It turns out that we can't put a copyMap in the board; so we get a
boardId for each installation, JSON.stringify the mapping of sha512 to
boardId, and put that in the board.