fix(amm): Prevent calls to remove Liquidity when the pool is empty
It's sufficient to prevent calls to remove liquidity that don't
include any liquidity tokens. If there are liquidity tokens
outstanding, then the pool won't be empty, and the divide-by-zero is
impossible.

The corresponding issue doesn't arise when adding liquidity, because
it explicity checks for an empty pool and uses the central token
amount to decide how much liquidity to create. All trading goes
through methods that ensure that the pools are non-empty and requested
amounts are positive.

fixes: #5131