fix(swingset): record full VatSyscallResult in transcript, even errors
Previously, the transcript only recorded the return data of syscalls,
and assumed that all syscalls succeeded (specifically it assumed that
a failed syscall would terminate the vat, thus removing the need for a
transcript).

But e.g. E(vatAdminSvc).getNamedBundleCap('missing') will provoke a
failing `syscall.invoke` to device-vat-admin, and we need to allow
vat-vat-admin to replay that correctly. So this changes the transcript
to record the full `VatSyscallResult` for all syscalls, including the
`ok`/`err` indicator, and the data of an error.

test-replay.js was changed to include a failed syscall in
vat-vat-admin. If the transcript patch is removed, this causes the
test to fail with "historical inaccuracy in replay of v2".

closes #5511