fix(xsnap): prohibit leading hyphen in options.name
The xsnap binary ignores argv entries that don't start with a hyphen,
and `xsnap.js` (ab)uses that to inject an extra no-op "name"
argument. This helps with debugging: by looking at `ps`, you can see
the argv arguments of each worker process, and figure out which worker
is hosting which vat.

To prevent a caller from using e.g. `options.name = '-l'` to inject
non-ignored xsnap options (changing a metering limit), this patch
asserts that `name` does not start with a hyphen.