fix(swingset): use gettimeofday() in the slogfile, not CLOCK_MONOTONIC
Unix provides multiple time APIs. `gettimeofday()` reports
seconds-since-epoch with microsecond resolution, and can jump around,
speed up, or even go backwards if/when someone changes the clock on
the host computer. `clock_gettime(CLOCK_MONOTONIC)` never goes
backwards, has nanosecond resolution, and in practice reports
seconds-since-last-reboot (of the host computer). `CLOCK_MONOTONIC` is
implemented (at least on linux) by initializing an offset (to the
current time) at boot, and reporting `gettimeofday() - offset`. If
someone calls `settimeofday()`, the kernel measures the delta, and
subtracts it from the offset, so the `CLOCK_MONOTONIC` value does
change even though the `gettimeofday()` does. However both
`gettimeofday()` and `CLOCK_MONOTONIC` will reflect the gentler
frequency adjustments that NTP makes (through `adjtime()`) to keep the
clock in sync without timequakes.

Previously, our slogfile's `.time` property recorded
`performance.timeOrigin + performance.now()`, which is effectively the
sum of `gettimeofday()` at process start, plus the delta in
`CLOCK_MONOTONIC` between one sample at process start and another when
`now()` is called, plus a tiny offset because `gettimeofday()` and the
initial sample cannot happen exactly simultaneously. This has the
benefit of never decreasing, and preserves the accuracy of
intervals (timestamp subtractions) that span a `settimeofday()`
timequake. But has the downside of becoming unique per-process when
timequakes happen: if you compare this value against a similar one
built in a different process, they will be different by the sum of any
timequakes that occurred between the first process launching and the
second one launching. If a process restarts, the timestamps before and
after the restart will be offset by a similar amount.

This changes the slogfile to use `gettimeofday()` for the `.time`
property, via a small C module named `microtime`. This matches what
issue #5152 changes in `xsnap-worker.c` to record delivery/syscall
send+receive timestamps. By using `gettimeofday()` on both sides, the
timestamps remain comparable independent of when the processes are
started or restarted. The drawback is that spans which cross a
timequake will be corrupted, even becoming negative if
`settimeofday()` causes the system clock to go backwards far
enough. However we only really pay attention to one span at a time (a
single block, or delivery, or syscall), so the damage is limited. I
care more about comparability of timestamps across long-running
processes (potentially across multiple hopefully-NTP-synced machines)
than I'm worried about the occasional timequake-corrupted span.

A new `.monotime` property records just `performance.now()`, which is
CLOCK_MONOTONIC relative to process start, and should have nanosecond
resolution (despite returning a number that absurdly claims
femtoseconds, probably because of base-2 representation limits). This
is good for higher-resolution interval computation, and is immune to
timequakes, but cannot be easily compared to anything outside the
process.

The slogfile lines were also rearranged, to put `type` at the front,
and the two timestamps at the end. This should make the file more
legible to humans who are just trying to eyeball the start/end of each
delivery.

If, some day, we manage to make swingset fast enough that the
microsecond resolution of `gettimeofday()` feels insufficient, we
should probably change both the slogger and `xsnap-worker.c` to use
the earlier approach: `timeOrigin + CLOCK_MONOTONIC` (which will
involve writing new code on the `xsnap-worker.c` side).

refs #5152