chore(xsnap): update to latest moddable XS version
The biggest change is a fix for Linux/ARM, where certain Promise
rejections were doing a fulfill by mistake.

I'm not sure if this changes the metering behavior, but there are
enough changes to the C code that I think it's likely. So I've also
bumped the metering version to 'xs-meter-15'.

closes #5587