fix: handle parallel requests to add an issuer in AMM (#5596)
* fix: handle parallel requests for the same issuer in AMM's addIssuer

refs: #5571

There are two bugs in #5571, and this fixes only one of them.

* chore: clarify a comment

* refactor: revert to multiple await style; handle race at .init()

Added a test for multiple attempts to add the same issuer with the
same keyword.

* refactor: rearrange error handling

drop try-catch for brand-to-promise

add catch for saving issuer and notifying Reserve

* chore: delete from brandToLiquidityMint on failure

name cleanup

* chore: cleanups: types, variable name, logging

* chore: move DISPLAY_INFO to top level scope

* chore: delete entries from brandToLiquidityIssuer when done

This is probably what was intended instead of the
code fixed in https://github.com/Agoric/agoric-sdk/pull/5611

* chore: convert a .then to E.when().

* refactor: return addIssuer as named (rather than anonymous) function